@@ -1,13 +1,12 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 25
+    compileSdkVersion 34
 
     defaultConfig {
-        minSdkVersion 9
-        targetSdkVersion 23
-        versionCode "git rev-list HEAD --count".execute().text.trim().toInteger()
-        versionName PROJ_VERSION
+        minSdkVersion 14
+        targetSdkVersion 34
+        versionCode 33
 
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
 
@@ -28,6 +27,50 @@ dependencies {
     androidTestImplementation 'com.android.support.test:runner:0.5'
 }
 
-if (hasProperty("bintrayUser" ) || hasProperty("BINTRAY_USER")) {
-    apply from: 'bintray.gradle'
+afterEvaluate {
+    publishing {
+        publications {
+            mavenJava(MavenPublication) {
+                groupId = "com.githang"
+                artifactId = "status-bar-compat"
+                version = "0.7"
+
+                // Include the AAR file manually
+                artifact("$buildDir/outputs/aar/${project.name}-release.aar") {
+                    builtBy tasks.named("assembleRelease")
+                }
+
+                pom {
+                    name = 'statusbarcompat'
+                    description = 'statusbarcompat'
+                    url = 'https://github.com/81813780/AVLoadingIndicatorView'
+                    licenses {
+                        license {
+                            name = 'The Apache License, Version 2.0'
+                            url = 'http://www.apache.org/licenses/LICENSE-2.0.txt'
+                        }
+                    }
+                    developers {
+                        developer {
+                            id = 'shay'
+                            name = 'Shay Brennan-Kelly'
+                            email = 'your.email@example.com'
+                        }
+                    }
+                }
+            }
+        }
+
+        repositories {
+            maven {
+                name = "GitHubPackages"
+                url = "https://maven.pkg.github.com/Shay-valid8me/status-bar-compat"
+                redentials {
+                    username = System.getenv("GITHUB_ACTOR") ?: "default_actor"
+                    password = System.getenv("GITHUB_TOKEN") ?: "default_token"
+                }
+            }
+        }
+    }
 }
+
