@@ -3,7 +3,7 @@ package com.githang.statusbar.demo;
 import android.app.Activity;
 import android.graphics.Color;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.webkit.WebView;
 
 import com.githang.statusbar.StatusBarCompat;
