@@ -5,8 +5,8 @@ import android.app.Activity;
 import android.graphics.Color;
 import android.os.Build;
 import android.os.Environment;
-import android.support.annotation.ColorInt;
-import android.support.annotation.IdRes;
+import androidx.annotation.ColorInt;
+import androidx.annotation.IdRes;
 import android.view.View;
 import android.view.ViewGroup;
 import android.view.Window;
